# Conflicts:
#	lib/bin/main/Gradle_Payroll/fileIO/Excel_Out.class

+++ lib/src/main/java/Gradle_Payroll/fileIO/Excel_Out.java
@@ -73,7 +73,7 @@ public class Excel_Out{
 		return 0;
 	}
 	
-	public static int formatRowHeight(int row,int height) {
+	public static int formatRowHeight(int row,float height) {
 		sheet.getRow(row).setHeightInPoints(height);
 		return 0;
 	}
